backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
pyodbc>=5.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
cryptography>=41.0.0
requests>=2.31.0
pandas>=2.0.0
pytz>=2024.1
python-multipart>=0.0.6
email-validator>=2.0.0
